requirements.txt
streamlit==1.32.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0